requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.3.1
Django==3.1.5
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
nodeenv==1.5.0
npm==0.1.1
optional-django==0.1.0
pycparser==2.20
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
